Salary Prediction using Linear Regression
Problem Statement 1

Your task is to help John from STA IT services to create a linear regression model in Python or R to predict the CTC/Salary of new hires from the data provided. Here is a snapshot of the case


Your task is to

1. Understand the data
2. Transform categorical variables to the numerical variables
3. Apply linear regression to predict the CTC
4. Evaluate the results of the analysis
5. Excel data files are provided to you to help you with the assignment.

Data_file.xlsx
Data_file is a dataset that you will use to train your model.

Test Data files:
Test_data_file.xlsx

Test_Data_file is a dataset that you will use to test your model and find out the Mean Square Value.